+++ host-configs/chaos_5_x86_64_ib/clang_4_0_0.cmake
@@ -1,7 +1,0 @@
-##############################################################################
-# Copyright (c) 2016-21, Lawrence Livermore National Security, LLC and Umpire
-# project contributors. See the COPYRIGHT file for details.
-#
-# SPDX-License-Identifier: (MIT)
-##############################################################################
-include(${CMAKE_CURRENT_LIST_DIR}/cudatoolkit_8_0.cmake)