+++ tools/replay/ReplayOptions.hpp
@@ -1,5 +1,5 @@
 //////////////////////////////////////////////////////////////////////////////
-// Copyright (c) 2016-21, Lawrence Livermore National Security, LLC and Umpire
+// Copyright (c) 2016-22, Lawrence Livermore National Security, LLC and Umpire
 // project contributors. See the COPYRIGHT file for details.
 //
 // SPDX-License-Identifier: (MIT)
@@ -35,6 +35,7 @@ struct ReplayUseHeuristicValidator : public CLI::Validator {
 };
 
 struct ReplayOptions {
+  ReplayOptions() {};
   bool time_replay_run{false};    // -t,--time-run
   bool time_replay_parse{false};  // --time-parse
   bool dump_statistics{false};    // -d, --dump